Extract functions mapping from t -> color into Theme
